+++ CHANGELOG.md
@@ -37,10 +37,10 @@ RELEASING:
 ### Added
 - migration from Junit4 to Junit5  ([#1320](https://github.com/GIScience/openrouteservice/pull/1320))
 - GTFS support for routing  ([#1316](https://github.com/GIScience/openrouteservice/pull/1316))
-- support for API time parameters  ([#1315](https://github.com/GIScience/openrouteservice/pull/1315))
 - upgrade to spring-boot 2.7.9 and dependency alignment for the api-tests [#1310](https://github.com/GIScience/openrouteservice/pull/1310)
 - moved all api tests into `openrouteservice` module [#1352](https://github.com/GIScience/openrouteservice/pull/1352)
 - migrate java from 11 to 17 [#1353](https://github.com/GIScience/openrouteservice/pull/1353)
+- time-dependent core routing algorithms ([#1388](https://github.com/GIScience/openrouteservice/pull/1388))
 
 ### Fixed
 - update docker setup description ([#1326](https://github.com/GIScience/openrouteservice/pull/1326))